Many new features und improvements

+++ backend/app/api/mailer.py
@@ -1,10 +1,12 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, Request
 from sqlalchemy.orm import Session
 from app.schemas.sender import SenderCreate, SenderOut, SenderUpdate
 from app.models.sender import Sender
 from app.services.smtp import test_smtp_connection
 from app.database import get_db
 from app.utils.responses import send_status_response
+
+from app.utils.security import Security
 
 router = APIRouter(prefix="/mailer", tags=["mailer"])
 
@@ -21,10 +23,12 @@ def test_sender(data: SenderCreate):
             detail=str(e)
         )
 
-@router.post("/", response_model=SenderOut)
-def create_sender(data: SenderCreate, db: Session = Depends(get_db)):
+@router.post("", response_model=SenderOut)
+def create_sender(request: Request, data: SenderCreate, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+
     try:
-        sender = Sender(**data.dict())
+        sender = Sender(**data.dict(), user_id=user.id)
         db.add(sender)
         db.commit()
         db.refresh(sender)
@@ -37,47 +41,83 @@ def create_sender(data: SenderCreate, db: Session = Depends(get_db)):
             detail=str(e)
         )
 
-@router.get("/", response_model=list[SenderOut])  # 👈 HIER
-def list_senders(db: Session = Depends(get_db)):
-    return db.query(Sender).all()
+@router.get("", response_model=list[SenderOut])  # 👈 HIER
+def list_senders(request: Request, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    return (db.query(Sender)
+            .filter(Sender.user_id == user.id)
+            .order_by(Sender.created_at.desc())
+            .all())
 
-@router.delete("/{sender_id}")
-def delete_webhook(sender_id: int, db: Session = Depends(get_db)):
-    sender = db.query(Sender).get(sender_id)
+@router.delete("/{mailer_id}")
+def delete_webhook(request: Request, mailer_id: str, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    sender = db.query(Sender).get(mailer_id)
+
+    if sender.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Cannot delete: unauthorized",
+            status=403,
+            detail=f"User {user.id} is not the owner of instance {mailer_id}."
+        )
+    
     if not sender:
         return send_status_response(
             code="DELETE_FAILED",
             message="Cannot delete: mailer not found",
             status=404,
-            detail=f"Mailer with id {sender_id} does not exist."
+            detail=f"Mailer with id {mailer_id} does not exist."
         )
+    
     db.delete(sender)
     db.commit()
     return {"success": True}
 
-@router.get("/{sender_id}", response_model=SenderOut)
-def get_sender(sender_id: int, db: Session = Depends(get_db)):
-    sender = db.query(Sender).get(sender_id)
+@router.get("/{mailer_id}", response_model=SenderOut)
+def get_sender(request: Request, mailer_id: str, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    sender = db.query(Sender).get(mailer_id)
+    
+    if sender.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Unauthorized access to sender",
+            status=403,
+            detail=f"User {user.id} is not allowed to access sender {mailer_id}."
+        )
+    
     if not sender:
         return send_status_response(
             code="MAILER_NOT_FOUND",
             message="Mailer not found",
             status=404,
-            detail=f"No mailer with id {sender_id} exists."
+            detail=f"No mailer with id {mailer_id} exists."
         )
+    
     return sender
 
-@router.put("/{sender_id}", response_model=SenderOut)
-def update_sender(sender_id: int, data: SenderUpdate, db: Session = Depends(get_db)):
-    sender = db.query(Sender).get(sender_id)
+@router.put("/{mailer_id}", response_model=SenderOut)
+def update_sender(request: Request, mailer_id: str, data: SenderUpdate, db: Session = Depends(get_db)):
+    user = Security(request).get_user()
+    sender = db.query(Sender).get(mailer_id)
+    
+    if sender.user_id != user.id:
+        return send_status_response(
+            code="UNAUTHORIZED",
+            message="Unauthorized access to sender",
+            status=403,
+            detail=f"User {user.id} is not allowed to update sender {mailer_id}."
+        )
+    
     if not sender:
         return send_status_response(
             code="UPDATE_FAILED",
             message="Cannot update: sender not found",
             status=404,
-            detail=f"Sender with id {sender_id} does not exist."
+            detail=f"Sender with id {mailer_id} does not exist."
         )
-
+    
     for key, value in data.dict(exclude_unset=True).items():
         setattr(sender, key, value)
